refactor: main app structure and implement service locator
