+++ lib/app/main_app.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../feature/theme/theme_cubit.dart';
+import 'router/app_router.dart';
+import 'service_locator/service_locator.dart';
+
+class MainApp extends StatelessWidget {
+  const MainApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => ThemeCubit(),
+      child: Builder(
+        builder: (context) {
+          final appRouter = sl<AppRouter>();
+          final themeMode = context.watch<ThemeCubit>().state;
+
+          return MaterialApp.router(
+            themeMode: themeMode,
+            theme: ThemeData.light(),
+            darkTheme: ThemeData.dark(),
+            title: "Health IA Coach",
+            routerConfig: appRouter.router,
+          );
+        },
+      ),
+    );
+  }
+}
